Move issuer test methods, reformat
