+++ aries_vcx/src/handlers/proof_presentation/mediated_prover.rs
@@ -1,10 +1,12 @@
-use crate::handlers::proof_presentation::prover::Prover;
-use crate::handlers::util::{matches_opt_thread_id, matches_thread_id};
-use crate::protocols::proof_presentation::prover::state_machine::ProverState;
+use std::collections::HashMap;
+
 use messages::msg_fields::protocols::notification::Notification;
 use messages::msg_fields::protocols::present_proof::PresentProof;
 use messages::AriesMessage;
-use std::collections::HashMap;
+
+use crate::handlers::proof_presentation::prover::Prover;
+use crate::handlers::util::{matches_opt_thread_id, matches_thread_id};
+use crate::protocols::proof_presentation::prover::state_machine::ProverState;
 
 #[allow(clippy::unwrap_used)]
 pub fn prover_find_message_to_handle(
