+++ aries_vcx/tests/utils/scenarios.rs
@@ -29,7 +29,12 @@ pub mod test_utils {
 
     use crate::utils::devsetup_alice::Alice;
     use crate::utils::devsetup_faber::Faber;
-    use crate::utils::devsetup_util::{get_credential_offer_messages, get_proof_request_messages, holder_update_with_mediator, prover_update_with_mediator};
+    use crate::utils::devsetup_util::get_credential_proposal_messages;
+    use crate::utils::devsetup_util::verifier_update_with_mediator;
+    use crate::utils::devsetup_util::{
+        get_credential_offer_messages, get_proof_request_messages, holder_update_with_mediator,
+        issuer_update_with_mediator, prover_update_with_mediator,
+    };
     use aries_vcx::common::ledger::transactions::into_did_doc;
     use aries_vcx::common::primitives::credential_definition::CredentialDef;
     use aries_vcx::common::primitives::revocation_registry::RevocationRegistry;
@@ -38,10 +43,6 @@ pub mod test_utils {
     use aries_vcx::handlers::connection::mediated_connection::{ConnectionState, MediatedConnection};
     use aries_vcx::handlers::issuance::holder::Holder;
     use aries_vcx::handlers::issuance::issuer::Issuer;
-    use aries_vcx::handlers::issuance::mediated_issuer::{
-        get_credential_proposal_messages, issuer_update_with_mediator,
-    };
-    use crate::utils::devsetup_util::verifier_update_with_mediator;
     use aries_vcx::handlers::proof_presentation::prover::Prover;
     use aries_vcx::handlers::proof_presentation::verifier::Verifier;
     use aries_vcx::protocols::issuance::holder::state_machine::HolderState;
